+++ docker/Dockerfile-simtools-prod
@@ -55,7 +55,6 @@ WORKDIR /workdir
 
 COPY LICENSE README.md pyproject.toml /workdir/simtools/
 COPY src /workdir/simtools/src
-COPY tools/generate_dependency_manifest.py /workdir/simtools/tools/generate_dependency_manifest.py
 
 ENV SIMTOOLS_SIM_TELARRAY_PATH="/workdir/simulation_software/sim_telarray"
 ENV SIMTOOLS_CORSIKA_PATH="/workdir/simulation_software/corsika7"
@@ -70,7 +69,7 @@ RUN python"${PYTHON_VERSION}" -m venv env \
   && . env/bin/activate \
   && SETUPTOOLS_SCM_PRETEND_VERSION_FOR_GAMMASIMTOOLS="${SIMTOOLS_VERSION}" \
     pip install --no-cache-dir /workdir/simtools \
-  && SIMTOOLS_CONTAINER_BUILD=1 python /workdir/simtools/tools/generate_dependency_manifest.py \
+  && SIMTOOLS_CONTAINER_BUILD=1 simtools-dependency-manifest \
   && pip cache purge && rm -rf /root/.cache \
   && echo ". /workdir/env/bin/activate" >> "/home/${CONTAINER_USER}/.bashrc" \
   && chown -R "${CONTAINER_UID}:${CONTAINER_GID}" /workdir "/home/${CONTAINER_USER}"
