+++ docs/source/api-reference/dependencies.md
@@ -38,6 +38,15 @@
    :members:
 ```
 
+(dependency-versions)=
+
+## dependency_versions
+
+```{eval-rst}
+.. automodule:: simtools.dependency_versions
+   :members:
+```
+
 (version)=
 
 ## version
