--- docs/source/user-guide/applications/simtools-dependency-versions.rst
@@ -1,6 +1,0 @@
-simtools-dependency-versions
-============================
-
-.. automodule:: simtools.applications.dependency_versions
-   :members:
-   :exclude-members: main
